requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
lxml==5.1.0
python-telegram-bot==13.15
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tornado==6.1
tzlocal==5.2
urllib3==2.2.0